Clean up pipeline build variables (#1722)
Remove the scripts that compute the BRANCH and CHANNEL at runtime.
Don't pass in the CHANNEL. This can be determined by the NETWORK instead (which is passed in, of course).
For now, compute the VERSION. This needs to be determined in a different manner and is a separate issue.
This also adds some improvements to the deploy scripts and sign script.

Although supporting command substitution is undoubtedly a good thing in the env blocks of the agents, this is made moot for now by assigning the ARCH_TYPE variable in the necessary dockerfiles and by simply doing the following to get the OS_TYPE:

OS_TYPE=$(uname)
OS_TYPE=${OS_TYPE,} # lowercases the first letter

Also, it turns out that the GOLANG_VERSION is not needed by the shell scripts, so this has been removed. If determined to be needed in the future, this can also be set in the Dockerfile (since both ARCH and GOLANG_VERSION are already docker build arguments).