Show all errors, not just the first, when compiling TEAL. (#1729)
* Show all errors, not just the first, when compiling TEAL.

Previously, we used the "usual" Go convention of returning an error to
the caller throughout assembly, as soon as they occured.  This meant
that assembly failed, as a whole, on the first bad line.  This model
also didn't work well for warnings, because they did not have the
context to do good reporting (they did not know the filename, for
example), but had no way to pass the wanring up to the caller who did
have the info (as errors did).  So this change also allows warnings to
report the file and line, for use by external tools.

The change is to accumulate errors and warnings in the OpStream
structure, rather than return immediately (or report immediately, iun
the case of wanrings).  After assembly the warnings and errors are
reported to stderr, and a summary error is used in Go to signal the
caller that assembly failed.

This allows external tools to use `goal clerk compile` to show all of
a script's problems at once.  I've added support to emacs in
teal-mode. We should consider a vscode plugin.