bugfix: incorrect block download retry accounting during fast catchup (#1792)
Incorrect block download retry accounting during fast catchup. The existing code was counting successful block retrieval as an retry attempt, which would "exhaust" the CatchupBlockDownloadRetryAttempts before the catchup is complete, causing the fast catchup to fail in case the CatchupBlockDownloadRetryAttempts was set too low.

The workaround of this is to set the CatchupBlockDownloadRetryAttempts to a value that is higher by 1000 than then number of desired attempts count.