simplify signed_last logic
The logic for signed_last was too complicated.  To simplify the logic,
this commit changes the block header from tracking CompactCertLastRound
(the last round for which we have a compact cert) to tracking
CompactCertNextRound instead (the next round for which we expect a
compact cert).

This subtle change allows distinguishing two cases that were previously
hard to tell apart: when compact certs are disabled (LastRound=0 and
NextRound=0) and when we need to sign the first compact cert after
enabling them in a consensus protocol upgrade (LastRound=0 but now
NextRound>0).

NextRound also precisely tells the signer worker thread where to start
signing.  If we already signed some block before stopping and re-starting,
duplicate signatures will be ignored.  If we deleted old ephemeral keys,
we will get an invalid signature, which will also be ignored.