Improvements/bugfixes for tealdbg (#1842)
Some transactions are not able to be fully expanded and viewed in tealdbg when using the Chrome DevTools frontend. This specifically affects transactions with empty `Accounts` or `ApplicationArgs` fields.

This bug happened because empty fields were being represented as nil slices in tealdbg, which translates to a `null` JSON value instead of an empty array, causing the Chrome frontend to fail to display any part of the expanded object. I've fixed this and made the following additional fixes in this PR:
* Made the object overflow field for Chrome more useful.
* Fixed an issue where the web frontend fails because the stack is empty.
* Fixed an issue where the web frontend fails to decode JSON with application local state because of integer map keys.

<details>
<summary>Current Chrome behavior (note the expanded arrow at gtxn[0] but nothing is shown)</summary>
<img src="https://user-images.githubusercontent.com/5856867/104783773-c5109600-5754-11eb-8f1b-88d733d45984.png">
</details>

<details>
<summary>Fixed Chrome behavior</summary>
<img src="https://user-images.githubusercontent.com/5856867/104783780-c8a41d00-5754-11eb-9ddf-6d5a2a72999d.png">
</details>