Change block header to allow for multiple compact certs in the future
This is a somewhat superficial change to how compact cert state appears in
the block header.  Instead of having the three fields (CompactCertVoters,
CompactCertVotersTotal, and CompactCertNextRound) live directly in the
block header, they now live in a map indexed by the type of compact cert.
There's just one type for now, but this change will enable us, in the
future, to cleanly support multiple types of compact certs.  (There might
be a type of compact cert for zero-knowledge proofs, or another type of
compact cert for PQ security.)

This change requires a bugfix to algorand-msgp, which I will file a PR for
shortly.