support Merkle commitment to transactions in block header
This will enable short proofs of transactions being present in a given
block.

The proof includes both the transaction ID (which is likely the most
useful part, knowing that some transaction ID was committed), as well as
the hash of the entire SignedTxnInBlock (which ensures all bits in the
Payset are committed to).

The Merkle tree uses the same merklearray code that's already used for
compact certificates.