faster computation of transaction ID and length
This change optimizes ID() and GetEncodedLength() for Transaction,
SignedTxn, and SignedTxnInBlock.  There are two sources of overhead that
this change avoids: interface conversions (to msgp.Marshaler) and memory
allocations (for temporary slices to hold the encoding).

To avoid interface conversions, the code invokes the msgp-generated
MarshalMsg() method.  To avoid error-checking boilerplate, the immediately
previous commit adds a msgp-generated MarshalMsgNoErr() method, which
panics on err exactly the same way that protocol.Encode() does.

To avoid memory allocation overhead, the code uses a sync.Pool to track
existing dirty slices that can be used for encoding.  This seems to be
more efficient than letting the Go GC manage slices, probably because
the encoder is OK with dirty (non-zero) slices.

This significantly improves performance according to BenchmarkEncoding
(in data/transactions).  Before this change:

  BenchmarkEncoding/EncodeLen-32          276892        4363 ns/op
  BenchmarkEncoding/ID-32                 169942        7131 ns/op

After this change:

  BenchmarkEncoding/EncodeLen-32         1963832         593 ns/op
  BenchmarkEncoding/ID-32                1109022        1042 ns/op