Remove data direcory from start/stop network expect tests. (#1902)
The expect tests were calling `goal network start -r <root dir> -d <data dir>` and `goal network stop -r <root dir> -d <data dir>`. While this is not harmful in any way, the neither of these commands is doing anything with the extra data directory parameter.

Given that this parameter is completely ignored, the is no point in passing it in. ( as it would only mislead the reader of these tests )