Remove GOPATH dependenncies in integration tests
Test has been GOPATH, but mostly just to find the top of the repor to
refer to tests.  This breaks if doing module oriented stuff instead of
GOPATH. These changes make it so that e2e.sh uses the full name of the
e2e test files, and e2e_client_runner executes the e2e scripts with
cwd set to top of repo.  That allowed removing the few places those
scripts used GOPATH, in favor of relative filenames.