Use total stake instead of online money
This commit will add `TotalStake` method to `LedgerReader` interface which returns the sum of voting stake of all accounts.

Currently `AccountData` struct has a `VotingStake` method, but there is no way for calculating the sum of voting stake of all accounts. We can only calculate the total money of online accounts. As a result, it's not possible to easily implement different agreement protocols by changing the behavior of `VotingStake` method.

Signed-off-by: aybehrouz <behrouz_ayati@yahoo.com>