Merge pull request #2019 from tsachiherman/tsachi/adjustoutgoingbufferssizes
Initial implementation of this code was using the worst-case single round message count as the number of pending outgoing messages. This was found to be more then just wasteful, but also problematic when it being used on relays.
From practical perspective, the agreement protocol for a node is only at a single step on any given moment. Therefore, creating a buffer that is big enough for that step should be enough. An exception for that are the late committees, which can run in parallel to the above step.

Accounting for both of the above reduces the practical outgoing buffer size from 38,000 to 11,000 - which in turns reduces the pending queues size from 3M -> 880K.