catchup: cross verify genesis.json hash (#2077)
When using the fast catchup, the block authenticity is verified against the catchpoint label. That method bypass the local genesis.json authentication since it's not required to establish secure catchup. From that point and on, the caught-up node would be able to proceed as usual. However, when the node is restarted, it compares the on-disk genesis.json to the first 320 rounds being loaded into memory. Failing with that comparison, the node would refuse to start up.

The solution here is to prevent the node from completing the fast catchup before testing that the genesis.json's hash matches the one on the first downloaded block. From there, it would be guaranteed to be identical due to the reverse hash check.