catchup: add deep block validation support (#2088)
Add a config key `CatchupBlockValidateMode` that would control which types of validation the catchup would perform prior to accept a block.

The default validation remains the same as before. However, it allows the end-user to remove/add certain validation "toggles". This enable us to add a deep validation support, that perform a complete validation of a block.