txnsync: improve outgoing message error handling (#2150)
Add error handling for outgoing messages:

If we fail to send a message to a peer, disconnect the peer ( it's probably already disconnected. But we can't tell that for sure ).
If the txnsync internally unable to enqueue the response for the successfully sent message, disconnect from peer. ( we could probably do better here, but that's an issue for future PR ).
In the emulator, detect txnsync error and panic on it.