txsync: avoid reevaluating old transactions during transaction selection (#2153)
This PR avoids reevaluating old transactions during transaction selection by maintaining the `lastTransactionSelectionGroupCounter`. This variable would get reset when new bloom filter arrives ( since we want to re-evaluate past transactions against the new bloom filter ),  as well as when the requested bloom filter changes.
In order to optimize that further, the transaction encoded size was moved to the transaction pool, ensuring that it would be executed only one time.