Ensure we stop kmd in case of a failuire to stop algod (#2169)
Existing implementation was not stopping the KMD in case we have failed to stop `algod`. This is wrong. A call to `NodeController.FullStop` should stop both of them, and return an error if it's unable to do so.

The concrete use case here was that algod has crashed while kmd remained in memory. Calling the above `FullStop()` failed since the `algod` was already gone, and therefore would not attempt to stop `kmd`.