testing: improve e2e protocol upgrade tests (#2176)
This PR addresses a bug in the protocol upgrade tests where the `AgreementFilterTimeout` was not set correctly.
In addition, this test removes `SendPaymentFromUnencryptedWallet()` and replaces it with separate `SendPaymentFromWallet()` and `GetUnencryptedWalletHandle()` calls for higher efficiency.

Last, this change *disables* the tests parallelism. Running the upgrade tests in parallel seems to be slower overall. This is still pending investigation, but can be decoupled from these changes for now.