Fix rekey upgrade test (#2178)
The existing rekey test was starting a network without rekeying support, making sure that rekeying transaction would fail there, and then waiting for the network to upgrade. Once the upgrade is complete, it would ensure that rekeying transactions works as expected.

The fail case in the above use case was that the upgrade would be asynchronous from the test itself, which was not taken into consideration. This change accounts for that and allow for slow execution of this test to take place and still pass, even if the network was advancing ahead of the test.