network: start network only after handlers registration is complete (#2177)
The various services used to register to the network callbacks as soon as they started. This was working mostly fine - but allowed a potential window where a connection could be established, and a message slip through. That message would not have a registered handler associated ( yet ), and therefore would be dropped.

The solution here is to flip the order : instead of starting the network first, and then starting all the rest of the component, we will start each of the components first, followed by starting the network. The current network implementation already knows how to provide buffering, and these would be more then adequate for the short period of time between the components starting and the network establishing it's connections.