More space efficient txnsync msg encoding (#2109)
Compress the TxnSync message before sending it and decompress it after receiving it. We can take advantage of the fact that some fields in transactions are duplicated as well as remove msgp tags by breaking arrays of txns into arrays of individual fields.

On a sample set of 20k txns on mainnet, this was able to encode the txnsync msg with ~37% less data.