combine app state key/value size limits (#2172)
Combine the limits enforced on keys and values in Teal such that there is a 128 byte limit on a key/value pair, rather than a 64 byte limit on each. The previous method meant app creators/users were essentially paying for ("renting") space that might be used by keys, but rarely was, and even if it was, could serve little purpose.