Regularize access to "foreign" references.
All opcodes that take accounts, asas, or apps can use the thing
itself (an address, or asa/app id) or an "indirect" reference through
the "foreign" arrays of the app transaction.  In all cases, the
"thing" must appear in those foreign arrays, it's just more convenient
sometimes to use the item in code, rather than indirect.