txnsync: optimize message encoding & decoding queue (#2279)
This PR optimize the message encoding and decoding by -

For decoding, it moves the decoding to be executed within the incoming message call from the network stack. This would give the mainloop thread more CPU time, and allow us to perform the incoming message decoding in parallel.
For encoding, we use the execution pool to asynchronously encode the message and send it to the network.
Use message buffers ( sync.Pool ) in order to avoid allocating memory for message data. This is particularly valuable for relays, which keeps allocating/releasing memory.