Fix docker build and push commands for releases
If you had a cached docker image, it would not build a new
one. Because tagging only references the tag to apply we just use the
updater script to fetch the latest for a channel, this meant we were
updating Docker Hub with pre-existing image builds.

This change will force a no-cache build. It also slightly refactors
some code for readability.