optimize txtail memory consumption (#2413)
The changes in this PR are as follows:

The unused method Ledger. GetRoundTxIds was removed. As a result, the txTail. getRoundTxIds can be removed as well. This makes the txids map stored in the roundTxMembers structure redundant.
In the cow.go, avoid adding empty leases to the cb.mods.Txleases map. Since we already not testing for empty leases, we can safely avoid storing them.
Optimize the txTail. loadFromDisk to generate optimal lastValid map sizes.
Optimize the txTail. loadFromDisk to avoid storing empty leases in the txleases map.