Add Extract opcodes  (#2521)
Adds extract opcodes that allow a substring to be extracted given a start index and a length (#2347).

The specs for the extract and extract3 opcodes are the same as the substring and substring3 opcodes, but takes a length rather than an end index. This also adds extract16bits, extract32bits, extract64bits which extracts 2, 4, and 8 byte strings respectively and converts them into integers.