Avoid requesting unsupported blocks by the ensure digest catchup path (#2563)
In current logic, the agreement is sending a notification to the catchup whenever it sees a cert vote
threshold without a corresponding block. When this happens on a round that is no longer supported
by the current binary, it would lead to continues requests for the said block. This PR addresses the issue
by testing if the coming block is expected to be supported or not. In case we know that it won't be supported,
we avoid downloading it entirely.