Infrastructure to check AVM stack manipulation opcodes better
This PR changes the way we check our understanding of the stack during
assembly.  We try to give errors as we assemble, based on the types
that we believe are on the stack when we assemble an opcode that has
stack expectations.  Previously, we *only* check the very top of the
stack, since almost all opcodes only looked at or modified the top. In
addition, we were very conservative with opcodes that manipulated the
stack.  For example, once we assembled a `swap` we only knew that the
top of the stack was [any, any]. If the top want [int, bytes], we lost
track.  Now we know the stack after that is [bytes, int].

Several opcodes still need this enhancement. "dig" and "swap" are
done, and serve as examples.  The following should be done next:
`dup`, `dup2`, `cover`, `uncover`.

Slightly trickier, we can also improve `select`.  If both of the
"then" and "else" types are the same, we can know that `select` leaves
that type behind.

To enhance an opcode, write a typeOpcode function that determines,
based on the opcodes immediates and on the current state of the stack,
what checkStack should do confirm the prior stack state (for example,
dig n needs a stack of depth n+1), and to setup the stack
afterward (for example, we know the stack after `swap` in more detail
that simply any,any.  Register this typeOpcode function using
"stacky()" in the opcodes table.