txnsync: avoid deadlock when disconnecting from network peer (#2603)
This PR address a deadlock between the `SendPeerMessage` and network package in the following scenario:
1. The `SendPeerMessage` is calling the peer `Unicast` method, passing a custom callback method.
2. The `Unicast` method returns nil, and the `SendPeerMessage` returns.
3. The network peer is failing to enqueue the message, making the call to the custom method specified in `SendPeerMessage`.
4. The custom method makes a call to the txsync, and concludes that the network connection need to be disconnected.
5. The custom method calls network's `Disconnect`, passing the peer.

At that point, the network package would block the `Disconnect` until the peer is disconnected. However, since the sending loop call is blocked due to the call to the callback method, the peer's `CloseAndWait` would never exit.

The solution implemented here is to return an error code from the callback method. When an error code is returned, the network package would treat that as a communication error and disconnect the peer.