fix TestVoteAggregatorBundles and check router.dispatch() output (#2619)
While working on the agreement package, I noticed the wrong messageEvent type was being used in TestVoteAggregatorBundles (voteVerified instead of a bundleVerified). This leads voteAggregator to use filterVote instead of filterBundle, and read the empty `messageEvent.Input.Vote` value (with round/period/step of 0/0/0) rather than the value of `messageEvent.Input.Bundle` that has the correct RPS and bundle of votes. This leads the events to be thrown away as from an old round instead of being processed by dispatch.