Fix for possible data race
TxHandler processDecodedArray and syncState evaluateIncomingMessage
may execute at the same time.

When this happens, evaluateIncomingMessage reads transactionGroups to
count the number if transactions there. At the same time,
processDecodedArray will be replacing verifiedTxGroup arrays with a
copy.

This leads to data race.

This change caches the number of transactions in the groups to avoid
reading transactionGroups to count that number.