txnsync: correctness e2e test (#2648)
This test sets up a network with 2 nodes and 2 relays.
The two nodes send payment transactions.
For each transaction, the test checks if the relays and the nodes (including the node that originated the transaction) have the transaction in the pool (i.e. the transactionInfo.ConfirmedRound == 0).

The tests needs to check for the transactions in the pool fast enough before they get evicted from the pool to the block.
To achieve this, it sends transactions during the first half of the round period, to give the test enough time to check for the transactions.